pyproject.toml
[project]
name = "ai-recipe-nutrition-app"
version = "1.0.0"
description = "AI-powered recipe and nutrition assistant"
authors = [{name = "Himesh Maniyar", email = "himeshmaniyar2017@gmail.com"}]
requires-python = ">=3.9"
dependencies = [
    "streamlit",
    "transformers>=4.21.0",
    "torch>=2.6.0",
    "torchvision>=0.17.0",
    "pillow",
    "requests",
    "safetensors",
    "beautifulsoup4",
    "faiss-cpu>=1.12.0",
    "sentence-transformers>=5.1.2",
    "numpy>=2.0.2",
]


